chg: introduced serialization package
