+++ src/controller/file/transmitting_to_fish.py
@@ -1,4 +1,4 @@
-from controller.file.write import create_xml
+from controller.file.serializing.general_serialization import create_xml
 from model import BoardConfiguration
 
 
@@ -8,7 +8,9 @@ def transmit_to_fish(show: BoardConfiguration, goto_default_scene: bool = True) 
         if scene.linked_bankset:
             if not scene.linked_bankset.is_linked:
                 scene.linked_bankset.link()
-    xml = create_xml(show)
+    xml = create_xml(show, assemble_for_fish_loading=True)
+    # TODO query current active scene
     show.broadcaster.load_show_file.emit(xml, goto_default_scene)
+    # TODO jump to current active scene and active bank set
     # TODO implement error handling
     return True
